+++ lib/widgets/total_score/total_score.dart
@@ -1,1 +1,2 @@
 export 'team_total_score_widget.dart';
+export 'total_score_row.dart';
