+++ lib/widgets/widgets.dart
@@ -1,2 +1,5 @@
+export 'app_bar/app_bar.dart';
 export 'game_info/game_info.dart';
+export 'scoreboard/scoreboard.dart';
+export 'scoreboard_screen_body.dart';
 export 'total_score/total_score.dart';
